chore: abstracts common openshift route locator logic.
