+++ openshift/openshift/src/main/java/org/arquillian/cube/openshift/impl/client/OpenShiftRouteLocator.java
@@ -1,60 +1,24 @@
-package org.arquillian.cube.openshift.impl.graphene.location;
+package org.arquillian.cube.openshift.impl.client;
 
 import io.fabric8.openshift.api.model.v3_1.Route;
 import org.arquillian.cube.kubernetes.api.Configuration;
-import org.arquillian.cube.openshift.impl.client.CubeOpenShiftConfiguration;
-import org.arquillian.cube.openshift.impl.client.OpenShiftClient;
 import org.jboss.arquillian.core.api.Instance;
-import org.jboss.arquillian.core.api.annotation.Inject;
-import org.jboss.arquillian.graphene.spi.configuration.GrapheneConfiguration;
-import org.jboss.arquillian.test.api.ArquillianResource;
-import org.jboss.arquillian.test.spi.enricher.resource.ResourceProvider;
 
-import java.lang.annotation.Annotation;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.Optional;
 
-import static org.arquillian.cube.openshift.impl.client.ResourceUtil.awaitRoute;
+public class OpenShiftRouteLocator {
 
-public class OpenshiftCustomizableURLResourceProvider implements ResourceProvider {
-
-    @Inject
-    private Instance<GrapheneConfiguration> grapheneConfiguration;
-
-    @Inject
     private Instance<OpenShiftClient> clientInstance;
-
-    @Inject
     private Instance<Configuration> configurationInstance;
 
-    @Override
-    public boolean canProvide(Class<?> type) {
-        return URL.class.isAssignableFrom(type);
+    public OpenShiftRouteLocator(Instance<OpenShiftClient> clientInstance, Instance<Configuration> configurationInstance) {
+        this.clientInstance = clientInstance;
+        this.configurationInstance = configurationInstance;
     }
 
-    @Override
-    public Object lookup(ArquillianResource resource, Annotation... qualifiers) {
-        try {
-            final URL routeUrl = resolveUrl();
-            awaitRoute(routeUrl);
-            return routeUrl;
-        } catch (MalformedURLException e) {
-            throw new IllegalArgumentException(e);
-        }
-    }
-
-    private URL resolveUrl() throws MalformedURLException {
-        final GrapheneConfiguration grapheneConfiguration = this.grapheneConfiguration.get();
-
-        final String configuredUrl = grapheneConfiguration.getUrl();
-        if (configuredUrl != null && !configuredUrl.isEmpty()) {
-            return getRoute(new URL(configuredUrl));
-        }
-        return getRoute();
-    }
-
-    private URL getRoute(URL routeUrl) {
+    public URL getRoute(URL routeUrl) {
         final CubeOpenShiftConfiguration config = getCubeOpenShiftConfiguration();
         final OpenShiftClient client = clientInstance.get();
 
@@ -66,7 +30,7 @@ public class OpenshiftCustomizableURLResourceProvider implements ResourceProvide
         return createUrlFromRoute(route);
     }
 
-    private URL getRoute() {
+    public URL getRoute() {
         final CubeOpenShiftConfiguration config = getCubeOpenShiftConfiguration();
         final OpenShiftClient client = clientInstance.get();
